requirements.txt
# Telegram bot
aiogram==3.4.1

# Database (PostgreSQL, async)
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Config / ENV
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
